address comments

+++ src/main/cpp/src/protobuf/protobuf.cu
@@ -910,11 +910,7 @@ std::unique_ptr<cudf::column> decode_protobuf_to_struct(cudf::column_view const&
               list_offsets, base_offset, d_locations.data(), i, num_scalar};
             auto valid_fn = [locs = d_locations.data(), i, num_scalar, has_def_str] __device__(
                               cudf::size_type row) {
-              return locs[flat_index(static_cast<size_t>(row),
-                                     static_cast<size_t>(num_scalar),
-                                     static_cast<size_t>(i))]
-                         .offset >= 0 ||
-                     has_def_str;
+              return locs[flat_index(row, num_scalar, i)].offset >= 0 || has_def_str;
             };
             column_map[schema_idx] = extract_and_build_string_or_bytes_column(false,
                                                                               message_data,
@@ -940,11 +936,7 @@ std::unique_ptr<cudf::column> decode_protobuf_to_struct(cudf::column_view const&
             list_offsets, base_offset, d_locations.data(), i, num_scalar};
           auto valid_fn = [locs = d_locations.data(), i, num_scalar, has_def_bytes] __device__(
                             cudf::size_type row) {
-            return locs[flat_index(static_cast<size_t>(row),
-                                   static_cast<size_t>(num_scalar),
-                                   static_cast<size_t>(i))]
-                       .offset >= 0 ||
-                   has_def_bytes;
+            return locs[flat_index(row, num_scalar, i)].offset >= 0 || has_def_bytes;
           };
           column_map[schema_idx] = extract_and_build_string_or_bytes_column(true,
                                                                             message_data,
